Move `frame_ticks()` and tick-type defs into `.ticktools`
